requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
chromadb>=0.4.0
openai>=1.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
